Move image download function to generic module
